requirements.txt
# requirements.txt

crewai==0.32.0
crewai-tools==0.2.6
langchain==0.1.20
duckduckgo-search-api